1. add thread pool benchmark test 2. optimize  MessageQueue::findInsertPositionLocked
on M2Pro release build

before
1-producers 1-workers time:1001ms,    839050 ops [   838211 ops/S] [  1193.02 ns/op]
1-producers 2-workers time:1000ms,   2779911 ops [  2779911 ops/S] [  359.724 ns/op]
1-producers 4-workers time:1000ms,   1114618 ops [  1114618 ops/S] [  897.168 ns/op]
2-producers 1-workers time:1001ms,    589169 ops [   588580 ops/S] [     1699 ns/op]
4-producers 1-workers time:1002ms,    424131 ops [   423284 ops/S] [  2362.48 ns/op]
2-producers 2-workers time:1000ms,   2995803 ops [  2995803 ops/S] [    333.8 ns/op]
4-producers 4-workers time:1000ms,   1292282 ops [  1292282 ops/S] [  773.825 ns/op]

after
1-producers 1-workers time:1000ms,   4298036 ops [  4298036 ops/s] [  232.664 ns/op]
1-producers 2-workers time:1000ms,   3203516 ops [  3203516 ops/S] [  312.157 ns/op]
1-producers 4-workers time:1000ms,   1712284 ops [  1712284 ops/S] [  584.015 ns/op]
2-producers 1-workers time:1000ms,   3951540 ops [  3951540 ops/S] [  253.066 ns/op]
4-producers 1-workers time:1000ms,   1818734 ops [  1818734 ops/S] [  549.833 ns/op]
2-producers 2-workers time:1000ms,   4836283 ops [  4836283 ops/S] [   206.77 ns/op]
4-producers 4-workers time:1000ms,   2185713 ops [  2185713 ops/S] [  457.517 ns/op]

+++ src/utils/MessageQueue.h
@@ -350,7 +350,7 @@ class MessageQueue {
    * \endcode
    *
    */
-  template <class Rep = int, class Period = std::milli>
+  template <class Rep = int64_t, class Period = std::milli>
   int32_t postMessage(const Message& message,
                       std::chrono::duration<Rep, Period> delay = std::chrono::milliseconds(0)) {
     using std::chrono::duration_cast;
